Refactor tests to put each type in it's own file.
This'll help when I add my tests, because otherwise this file will explode in size
